+++ src/jobs/ImportJob.php
@@ -6,9 +6,12 @@
 namespace putyourlightson\campaign\jobs;
 
 use Craft;
+use craft\helpers\Queue;
 use craft\queue\BaseJob;
+use craft\queue\jobs\UpdateSearchIndex;
 use DateTime;
 use putyourlightson\campaign\Campaign;
+use putyourlightson\campaign\elements\ContactElement;
 use putyourlightson\campaign\events\ImportEvent;
 use putyourlightson\campaign\services\ImportsService;
 
@@ -60,6 +63,9 @@ class ImportJob extends BaseJob
 
         // Save import
         Campaign::$plugin->imports->saveImport($import);
+
+        // Update the search indexes
+        Campaign::$plugin->imports->updateSearchIndexes();
 
         // Fire an after event
         if (Campaign::$plugin->imports->hasEventHandlers(ImportsService::EVENT_AFTER_IMPORT)) {